requirements.txt
PySimpleGUI == 4.60.5
tkinterdnd2 == 0.3.0
Pillow == 2.2.1